Moving search classes to right packages
