+++ search/src/main/res/transition/search_hide_confirm.xml
@@ -21,7 +21,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <transition
-        class="io.plaidapp.ui.search.transitions.CircularReveal"
+        class="io.plaidapp.search.ui.transitions.CircularReveal"
         app:endRadius="@dimen/fab_radius"
         android:duration="150"
         android:interpolator="@android:interpolator/fast_out_slow_in">
