+++ search/src/main/res/transition/search_return.xml
@@ -19,7 +19,7 @@
 <transitionSet xmlns:android="http://schemas.android.com/apk/res/android">
 
     <transition
-        class="io.plaidapp.ui.search.transitions.CircularReveal"
+        class="io.plaidapp.search.ui.transitions.CircularReveal"
         android:duration="600"
         android:interpolator="@android:interpolator/fast_out_linear_in">
         <targets>
